+++ src/services/auth/oauth2/types.ts
@@ -7,6 +7,7 @@ import {
 
 export interface AuthOauth2Service {
   tokens: AuthOauth2TokenService;
+  refreshTokens: OAuth2RefreshTokenService;
 
   /**
    * Create an OAuth2 authorization
@@ -22,14 +23,15 @@ export interface AuthOauth2Service {
   createAuthorization(
     data: OAuth2AuthorizationCreation,
     options?: OptionsBase
-  ): Promise<OAuth2Authorization>;
+  ): Promise<OAuth2AuthorizationCreationResponse>;
 
   /**
    * Get a list of OAuth2 Authorizations
    *
    * Permission | Scope | Effect
    * - | - | -
-   * VIEW_AUTHORIZATIONS | global | **Required** for this endpoint
+   * none | | Can only see a list of OAuth2 authorizations for this account
+   * VIEW_AUTHORIZATIONS | global | See any authorizations belonging to any user
    * @see https://swagger.extrahorizon.com/swagger-ui/?url=https://swagger.extrahorizon.com/auth-service/2.0.4-dev/openapi.yaml#/OAuth2/get_oauth2_authorizations
    */
   getAuthorizations(
@@ -41,7 +43,8 @@ export interface AuthOauth2Service {
    *
    * Permission | Scope | Effect
    * - | - | -
-   * DELETE_AUTHORIZATIONS | global | **Required** for this endpoint
+   * none | | Can only delete OAuth2 authorizations for this account
+   * DELETE_AUTHORIZATIONS | global | Delete any authorizations belonging to any user
    * @see https://swagger.extrahorizon.com/swagger-ui/?url=https://swagger.extrahorizon.com/auth-service/2.0.4-dev/openapi.yaml#/OAuth2/delete_oauth2_authorizations__authorizationId_
    * @throws {ResourceUnknownError}
    */
@@ -58,7 +61,7 @@ export interface AuthOauth2TokenService {
    * Permission | Scope | Effect
    * - | - | -
    * none | | Can only see a list of OAuth2 tokens for this account
-   * VIEW_AUTHORIZATIONS | global | Can see a list of OAuth2 tokens for any account
+   * VIEW_AUTHORIZATIONS | global | See any OAuth2 tokens belonging to any user
    */
   find(options?: OptionsWithRql): Promise<PagedResult<OAuth2Token>>;
 
@@ -68,7 +71,7 @@ export interface AuthOauth2TokenService {
    * Permission | Scope | Effect
    * - | - | -
    * none | | Can only see a list of OAuth2 tokens for this account
-   * VIEW_AUTHORIZATIONS | global | Can see a list of OAuth2 tokens for any account
+   * VIEW_AUTHORIZATIONS | global | See any OAuth2 tokens belonging to any user
    */
   findAll(options?: OptionsWithRql): Promise<OAuth2Token[]>;
 
@@ -78,7 +81,7 @@ export interface AuthOauth2TokenService {
    * Permission | Scope | Effect
    * - | - | -
    * none | | Can only see a list of OAuth2 tokens for this account
-   * VIEW_AUTHORIZATIONS | global | Can see a list of OAuth2 tokens for any account
+   * VIEW_AUTHORIZATIONS | global | See any OAuth2 tokens belonging to any user
    */
   findFirst(options?: OptionsWithRql): Promise<OAuth2Token | undefined>;
 
@@ -88,7 +91,7 @@ export interface AuthOauth2TokenService {
    * Permission | Scope | Effect
    * - | - | -
    * none | | Can only see a list of OAuth2 tokens for this account
-   * VIEW_AUTHORIZATIONS | global | Can see a list of OAuth2 tokens for any account
+   * VIEW_AUTHORIZATIONS | global | See any OAuth2 tokens belonging to any user
    */
   findById(id: string, options?: OptionsWithRql): Promise<OAuth2Token | undefined>;
 
@@ -97,29 +100,121 @@ export interface AuthOauth2TokenService {
    *
    * Permission | Scope | Effect
    * - | - | -
-   * DELETE_AUTHORIZATIONS | | Required for this endpoint
+   * none | | Can only delete OAuth2 tokens for this account
+   * DELETE_AUTHORIZATIONS | global | Delete any OAuth2 tokens belonging to any user
    */
   remove(id: string): Promise<AffectedRecords>;
+}
+
+export interface OAuth2RefreshTokenService {
+  /**
+   * # Get a list of OAuth2 refresh tokens
+   *
+   * Permission | Scope | Effect
+   * - | - | -
+   * none | | Can only see a list of OAuth2 refresh tokens for this account
+   * VIEW_OAUTH2_REFRESH_TOKENS or VIEW_AUTHORIZATIONS | global | Can see a list of OAuth2 refresh tokens for any account
+   * Using VIEW_AUTHORIZATIONS for this endpoint is deprecated; use VIEW_OAUTH2_REFRESH_TOKENS instead
+   * @param options.rql Add filters to the requested list
+   * @returns PagedResult<OAuth2RefreshToken>
+   */
+  find(options?: OptionsWithRql): Promise<PagedResult<OAuth2RefreshToken>>;
+
+  /**
+   * Get a list of OAuth2 refresh tokens
+   *
+   * Permission | Scope | Effect
+   * - | - | -
+   * none | | Can only see a list of OAuth2 refresh tokens for this account
+   * VIEW_OAUTH2_REFRESH_TOKENS or VIEW_AUTHORIZATIONS | global | Can see a list of OAuth2 refresh tokens for any account
+   * Using VIEW_AUTHORIZATIONS for this endpoint is deprecated; use VIEW_OAUTH2_REFRESH_TOKENS instead
+   * @param options.rql Add filters to the requested list
+   * @returns OAuth2RefreshToken[]
+   */
+  findAll(options?: OptionsWithRql): Promise<OAuth2RefreshToken[]>;
+
+  /**
+   * Get the first OAuth2 refresh token found
+   *
+   * Permission | Scope | Effect
+   * - | - | -
+   * none | | Can only see a list of OAuth2 refresh tokens for this account
+   * VIEW_OAUTH2_REFRESH_TOKENS or VIEW_AUTHORIZATIONS | global | Can see a list of OAuth2 refresh tokens for any account
+   * Using VIEW_AUTHORIZATIONS for this endpoint is deprecated; use VIEW_OAUTH2_REFRESH_TOKENS instead
+   * @param options.rql Add filters to the requested list
+   * @returns {Promise<OAuth2RefreshToken | undefined>}
+   */
+  findFirst(options?: OptionsWithRql): Promise<OAuth2RefreshToken | undefined>;
+
+  /**
+   * Get an oAuth2 refresh token by its id
+   *
+   * Permission | Scope | Effect
+   * - | - | -
+   * none | | Can only see a list of OAuth2 refresh tokens for this account
+   * VIEW_OAUTH2_REFRESH_TOKENS or VIEW_AUTHORIZATIONS | global | Can see a list of OAuth2 refresh tokens for any account
+   * Using VIEW_AUTHORIZATIONS for this endpoint is deprecated; use VIEW_OAUTH2_REFRESH_TOKENS instead
+   * @param id the refresh token id
+   * @param options.rql Add filters to the requested list
+   * @returns {Promise<OAuth2RefreshToken | undefined>}
+   */
+  findById(id: string, options?: OptionsWithRql): Promise<OAuth2RefreshToken | undefined>;
+
+  /**
+   * Delete an oAuth2 refresh token
+   *
+   * Permission | Scope | Effect
+   * - | - | -
+   * none | | Can only delete OAuth2 refresh tokens for this account
+   * DELETE_OAUTH2_REFRESH_TOKEN or DELETE_AUTHORIZATIONS | global | Delete any OAuth2 refresh tokens belonging to any user
+   * Using DELETE_AUTHORIZATIONS for this endpoint is deprecated; use DELETE_OAUTH2_REFRESH_TOKEN instead
+   * @param id the refresh token id
+   * @returns AffectedRecords
+   */
+  remove(id: string, options?: OptionsBase): Promise<AffectedRecords>;
 }
 
 export interface OAuth2AuthorizationCreation {
   responseType: string;
   clientId: string;
+  redirectUri?: string;
+  state?: string;
+  codeChallengeMethod?: string;
+  codeChallenge?: string;
+}
+
+export interface OAuth2AuthorizationCreationResponse {
+  id: string;
+  clientId: string;
+  userId: string;
   redirectUri: string;
-  state: string;
-  scope: string;
+  state?: string;
+  codeChallengeMethod?: string;
+  codeChallenge?: string;
+  authorizationCode: string;
+  expiryTimestamp: Date;
+  updateTimestamp: Date;
+  creationTimestamp: Date;
 }
 
 export interface OAuth2Authorization {
   id: string;
-  userId: string;
   clientId: string;
-  authorizationCode: string;
-  state: string;
-  /** The timestamp when the authorization was last updated */
-  updateTimestamp?: Date;
-  /** The timestamp when the authorization was created */
-  creationTimestamp?: Date;
+  userId: string;
+  redirectUri: string;
+  state?: string;
+  codeChallengeMethod?: string;
+  /**
+   @deprecated `codeChallenge` will be removed from responses returned by listing endpoints in a future version.
+   */
+  codeChallenge?: string;
+  /**
+   * @deprecated `authorizationCode` will be removed from responses returned by listing endpoints in a future version.
+   */
+  authorizationCode?: string;
+  expiryTimestamp: Date;
+  updateTimestamp: Date;
+  creationTimestamp: Date;
 }
 
 export interface OAuth2Token {
@@ -127,7 +222,19 @@ export interface OAuth2Token {
   applicationId: string;
   userId: string;
   refreshTokenId: string;
-  accessToken: string;
+  /**
+   * @deprecated `accessToken` will be removed from responses returned by listing endpoints in a future version.
+   */
+  accessToken?: string;
+  expiryTimestamp: Date;
+  updateTimestamp: Date;
+  creationTimestamp: Date;
+}
+
+export interface OAuth2RefreshToken {
+  id: string;
+  applicationId: string;
+  userId: string;
   expiryTimestamp: Date;
   updateTimestamp: Date;
   creationTimestamp: Date;