+++ CHANGELOG.md
@@ -7,6 +7,28 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [8.10.0]
 
+### Added
+- `exh.users.getEmailTemplates` and `exh.users.setEmailTemplates` now support `_name` variants
+  - `activationEmailTemplateName`
+  - `reactivationEmailTemplateName`
+  - `passwordResetEmailTemplateName`
+  - `oidcUnlinkEmailTemplateName`
+  - `oidcUnlinkPinEmailTemplateName`
+  - `activationPinEmailTemplateName`
+  - `reactivationPinEmailTemplateName`
+  - `passwordResetPinEmailTemplateName`
+
+### Deprecated
+- `exh.users.getEmailTemplates` and `exh.users.setEmailTemplates` deprecates `_id` in favour of `_name` variants
+  - `activationEmailTemplateId`
+  - `reactivationEmailTemplateId` 
+  - `passwordResetEmailTemplateId`
+  - `oidcUnlinkEmailTemplateId`
+  - `oidcUnlinkPinEmailTemplateId`
+  - `activationPinEmailTemplateId`
+  - `reactivationPinEmailTemplateId`
+  - `passwordResetPinEmailTemplateId`
+
 ## [8.9.1]
 
 ### Fixed